+++ server/RoomManager.hpp
@@ -9,7 +9,7 @@ class RoomManager{
         std::set<int> client_id_set;
         int host_id;
         int id;
-        unsigned int map_seed;
+        int map_seed;
         inline int player_count() const;
         inline bool is_full() const;
     };
@@ -21,11 +21,11 @@ public:
     bool join_room(const int client_id, int &room_id);
     //return true if host change
     bool exit_room(const int, const int);
-    void set_map_seed(const int, const unsigned long);
+    void set_map_seed(const int, const int);
     //start create threads for the game and change state
     void start_game(const int);
     int get_host_id(const int) const;
-    const unsigned long get_map_seed(const int) const;
+    const int get_map_seed(const int) const;
     int player_count(const int) const;
     const std::set<int>& get_clients(const int) const;
     void check_state();
